backend/pyproject.toml
[project]
name = "athira-availability-api"
version = "0.1.0"
description = "Athira tutor availability simulation API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.6.0",
  "sqlalchemy>=2.0.36",
  "httpx>=0.27.0",
  "google-generativeai>=0.8.0",
  "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
